Move data structures to jaxsim.api.ode_data
